requirements.txt
aniso8601==8.1.1
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-graphql-jwt==0.3.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
graphene==3.0b7
graphene-django==3.0.0b7
graphql-core==3.1.4
graphql-relay==3.1.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
pipenv==2020.11.15
promise==2.3
pycparser==2.20
PyJWT==2.0.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
virtualenv==20.4.3
virtualenv-clone==0.5.4
python-dotenv==0.19.2
pycryptodome==3.11.0

